block: bio: pass bvec table to bio_init()
Some drivers often use external bvec table, so introduce
this helper for this case. It is always safe to access the
bio->bi_io_vec in this way for this case.

After converting to this usage, it will becomes a bit easier
to evaluate the remaining direct access to bio->bi_io_vec,
so it can help to prepare for the following multipage bvec
support.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>

Fixed up the new O_DIRECT cases.

Signed-off-by: Jens Axboe <axboe@fb.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>

# Conflicts:
#	drivers/scsi/ufs/ufshpb.c
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: TheSync <repo-sync@outlook.com>