binfmt_elf: Don't allocate memory dynamically in load_elf_binary
The dynamic memory allocations in load_elf_binary can be replaced by
large stack allocations that bring the total frame size for the function
up to about 4 KiB. We have 16 KiB of stack space, and this function does
not have deep function call depths, so such a large stack allocation is
safe. This improves performance by eliminating the overhead of dynamic
memory allocation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: Iae59d6b577392b6093732e87a5b1b32cdbd6f63d
Signed-off-by: Hadad <repo-sync@outlook.com>