sched/fair.c: Bring back the original intent of sync_boost
Before this change, sync_boost would only boost if the task was
already boosted. Thus making it do nothing.

The original intent:

If the sync flag is ignored because the current cpu is not in
the affinity mask for the target of a sync wakeup (usually
binder call), prefer to place in the same cluster if
possible. The main case is a "top-app" task waking a
"foreground" task when the top-app task is running on
a CPU that is not in the foreground cpuset. This patch
causes the "top-app" search order be used (big cpus first)
even if schedtune.boost is 0 when the sync flag failed.

Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: TheSync <repo-sync@outlook.com>