techpack: audio: Add support for legacy Q6 service versions
Instead of throwing an error everytime and falling back to old
service initialization and handling, in case the DSP firmware
does NOT return the services versions, suppose that we are
dealing with legacy DSP firmwares on legacy SoCs and also that
the all the supported legacy SoCs have got V1 service versions.

This is true and usable for a number of legacy SoCs, like
MSM8956/76, 8996, 8998, SDM630/636/660.

Signed-off-by: TheSync <repo-sync@outlook.com>